+++ server/index.js
@@ -1,7 +1,22 @@
+const fs = require('fs');
+const path = require('path');
+
 // Transpile all code following this line with babel and use 'env' (aka ES6) preset.
 require('babel-register')({
   presets: [ 'env' ]
-})
+});
 
-// Import the rest of our application.
-module.exports = require('./server.js')
+const socket = require('./socket');
+const mongo = require('./mongo');
+
+const options = {
+  key: fs.readFileSync(path.resolve(__dirname, '../sample/cert.pem')),
+  cert: fs.readFileSync(path.resolve(__dirname, '../sample/cert.crt'))
+};
+const port = process.argv[3] || 3000;
+
+const url = `mongodb://127.0.0.1:27017/`
+
+mongo.enable(url, () => {
+  socket.enable(options, port);
+});
